hwbench/monitoring: Add types in monitoring.cfg
The monitoring config file is fully dedicated to the BMC.

This commit is about adding a "type" to each section to let room for
future monitoring components like PDU, CDU etc..

This commit is adding "type=BMC" for the current BMC examples.
The BMC code is also moved out of vendor.py to make it easier to read.
