+++ monitoring.cfg
@@ -5,11 +5,14 @@
 #[HPE]
 #username=Administrator
 #password=YOURPASSWORD
+#type=BMC
 
 #[DELL]
 #username=Administrator
 #password=YOURPASSWORD
+#type=BMC
 
 [default]
 username=Administrator
 password=YOURPASSWORD
+type=BMC
